% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot crash locations}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The state for which accident locations are to be plotted, specified as a number.}

\item{year}{The 4-digit year for which the data should be plotted}
}
\value{
A simple plot of accident locations.
}
\description{
This function generates a simple plot of the locations of traffic accidents
for a specified state. Errors are thrown if an invalid year or state number is input,
and a warning is generated if there are no accidents for the year/state specified.
}
\examples{
fars_map_state(1, 2014) # Alabama map from 2014
fars_map_state(18, 2013) # Indiana map from 2013

}
