% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read fars data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The name of the file to be imported}
}
\value{
This function returns a tibble of the set specified.
}
\description{
This function reads a fars data set from the working directory. It first checks
to make sure the specified file actually exists, then imports it with read_csv
if it does. An error is thrown if the file does not exist.
}
\examples{
\dontrun{
fars_data <- readr::read_csv("accident_2013.csv.bz2")
}

}
