% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read multiple fars data sets simultaneously}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{An integer vector of years.}
}
\value{
A list of tibbles, one per year specified. Entries in this list will
be null if corresponding values of the input do not give rise to valid file names.
}
\description{
This functions accepts a vector of years and returns a list of fars data frames.
A warning message is passed if any of the years specified do not correspond to
valid files in the workingn directory.
}
\examples{
years <- 2013:2015
fars_list <- fars_read_years(years)

}
