% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Import and summarize multiple fars data sets.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A numeric vector of years.}
}
\value{
A tibble with 12 rows, one for each calendar month. In addition to a column
of months, the output includes one column for each year specified.
}
\description{
This function imports multiple fars data sets, specified by a numeric vector of
years, and returns a tibble with monthly counts for those years.
}
\examples{
\dontrun{
years <- 2013:2015
fars_summary <- fars_summarize_years(years)
}


}
