% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Build a fars file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year of the data set. Must be an integer or coercible to one.}
}
\value{
This function returns a file path in the standard format of compressed fars files.
}
\description{
This function builds a file name for a compressed fars data set using only the
year of the set. Note that no checks are run to ensure that a file with
the given name actually exists in the working directory.
}
\examples{
file_path <- make_filename(2014)
}
